requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
requests
streamlit
pandas
plotly
aiosqlite
